CATALOG-1651: sessions add register remote database engine
