+++ query/src/sessions/sessions.rs
@@ -32,6 +32,7 @@ use crate::catalogs::Catalog;
 use crate::clusters::ClusterRef;
 use crate::configs::Config;
 use crate::datasources::local::LocalDatabases;
+use crate::datasources::remote::RemoteDatabases;
 use crate::datasources::system::SystemDatabases;
 use crate::sessions::session::Session;
 use crate::sessions::session_ref::SessionRef;
@@ -50,11 +51,11 @@ pub type SessionManagerRef = Arc<SessionManager>;
 impl SessionManager {
     pub fn from_conf(conf: Config, cluster: ClusterRef) -> Result<SessionManagerRef> {
         let catalog = Arc::new(DatabaseCatalog::try_create_with_config(conf.clone())?);
-        // Register local database engine.
-        catalog.register_db_engine("local", Arc::new(LocalDatabases::create()))?;
-        // Register system database engine.
-        catalog.register_db_engine("system", Arc::new(SystemDatabases::create()))?;
-        // TODO(bohu): register remote database engine.
+
+        // Register local/system and remote database engine.
+        catalog.register_db_engine("local", Arc::new(LocalDatabases::create(conf.clone())))?;
+        catalog.register_db_engine("system", Arc::new(SystemDatabases::create(conf.clone())))?;
+        catalog.register_db_engine("remote", Arc::new(RemoteDatabases::create(conf.clone())))?;
 
         let max_active_sessions = conf.query.max_active_sessions as usize;
         Ok(Arc::new(SessionManager {
